requirements.txt
beautifulsoup4==4.6.3
dj-database-url==0.5.0
Django==2.1.5
django-ckeditor==5.6.1
django-contrib-comments==1.9.0
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-pagedown==1.0.6
django-summernote==0.8.11.4
django-threadedcomments==1.0b1
djangorestframework==3.9.1
gunicorn==19.9.0
Jinja2==2.10.1
jsonschema==3.0.1
nltk==3.4.5
pyOpenSSL==19.0.0
qtconsole==4.4.3
queuelib==1.5.0
reportlab==3.3.0
requests==2.20.1
scipy==1.1.0
Scrapy==1.7.3
Unidecode==1.1.1
virtualenv==16.4.3
whitenoise==4.1.2


tensorflow==1.15.0
pandas==0.25.3
Pillow==6.2.1
numpy==1.17.4
opencv-python==4.1.1.26
Keras==2.2.5
h5py==2.10.0